qwen3a/requirements.txt
transformers>=4.41.0
accelerate>=1.10.0
torchaudio
soundfile
datasets
peft>=0.10.0
sentencepiece
numpy
scipy
bitsandbytes
